feat: change landing page side bar
